Refactor DRI_PRIME detection and allow child classes
Previously, the DRI_PRIME detection code was very monolithic and a bit
messy.

With the advent of EGL and Wayland/ANGLE the need for another
implementation arose and thus an already complicated method would've
required further duplication.

As a solution, a new `PrimeLinuxBSD` class has been created with a
public big generic method implemented and some backend-specific virtual
methods for initializing the graphics context and accessing the graphics
library.

Note that the `detect_prime` (now `detect_gpu`, read below) is not
static anymore to allow the virtual methods to be called. This means
that the class has now to be instantiated for the method to be called
but I couldn't find a cleaner method (and AFAIU this whole thing is
partly a hack anyway)

Also, while I was at it, the big probing method got a few changes that
will hopefully improve readibility and improve it a little bit:

 - All the probed device data has been moved to a struct instead of
using individual arrays for each property and the struct members
initialization has been moved to its definition.

 - The `detect_prime` method has been renamed to `detect_gpu` as the
actual "PRIME" identification is done in the display server. I'd
actually also have added "index" or something like that but it would've
made the name too long.

 - The amount of device probes has been increased to 4 due to some users
which have also a software rasterizer not being able to detect the
actual dedicated GPU.

 - Some more "modern" goodies have been used such as `vformat` and the
`for` iterator notation.

That said, I also kept untouched some more more complex parts, such as
the tricky IPC logic and the priority calculation, as I don't want to
change behavior further on an already relatively delicate piece of code.